+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - When generating new mnemonic, show it without a trace.
 - Support import/export slashing protection interchange data [EIP-3076](https://eips.ethereum.org/EIPS/eip-3076)
+- Support the new command `run`, used to run all the services generated
 
 ### Changed